backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
openai==1.57.2
langchain==0.3.9
langchain-community==0.3.9
langchain-openai==0.2.11
faiss-cpu==1.13.0
sentence-transformers==3.3.1
tiktoken==0.8.0
httpx==0.28.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
twilio==9.4.0